change: separate CompetenceTypePage and CompetenceTypeEditor
