+++ src/pages/Settings/CompetencesPage/CompetencesSettingsPage.tsx
@@ -1,14 +1,13 @@
-import { type MouseEvent, useEffect, useState } from "react";
-import { useLoaderData, useNavigate } from "react-router-dom"
+import { useState } from "react";
+import { useLoaderData, useNavigate } from "react-router-dom";
 
 import UniversimeApi from "@/services/UniversimeApi";
 import { Filter } from "@/components/Filter/Filter";
-import { type CompetencesSettingsLoaderResponse, SettingsDescription, SettingsTitle } from "@/pages/Settings";
+import { type CompetencesSettingsLoaderResponse, SettingsDescription, SettingsTitle, CompetenceTypeEditor } from "@/pages/Settings";
 
 import * as SwalUtils from "@/utils/sweetalertUtils";
 import { stringIncludesIgnoreCase } from "@/utils/stringUtils";
 import { groupArray } from "@/utils/arrayUtils";
-import { makeClassName, setStateAsValue } from "@/utils/tsxUtils";
 
 import { type CompetenceType } from "@/types/Competence";
 import "./CompetencesSettingsPage.less";
@@ -65,7 +64,7 @@ export function CompetencesSettingsPage() {
 
             <div className="competences-wrapper">
             { unreviewedFilteredCompetences.length > 0
-                ? unreviewedFilteredCompetences.map(ct => <CompetenceTypeEditor key={ct.id} ct={ct} refreshCTs={refreshCompetenceTypes} />)
+                ? unreviewedFilteredCompetences.map(ct => <CompetenceTypeEditor key={ct.id} ct={ct} refreshCompetenceTypes={refreshCompetenceTypes} />)
                 : <p className="empty-competences">
                 { unreviewedCompetenceTypes.length > 0
                     ? "Nenhuma competência encontrada."
@@ -88,7 +87,7 @@ export function CompetencesSettingsPage() {
 
             <div className="competences-wrapper">
             { reviewedFilteredCompetences.length > 0
-                ? reviewedFilteredCompetences.map(ct => <CompetenceTypeEditor key={ct.id} ct={ct} refreshCTs={refreshCompetenceTypes} />)
+                ? reviewedFilteredCompetences.map(ct => <CompetenceTypeEditor key={ct.id} ct={ct} refreshCompetenceTypes={refreshCompetenceTypes} />)
                 : <p className="empty-competences">
                 { reviewedCompetenceTypes.length > 0
                     ? "Nenhuma competência encontrada."
@@ -103,114 +102,5 @@ export function CompetencesSettingsPage() {
     async function refreshCompetenceTypes() {
         const competenceTypes = await UniversimeApi.CompetenceType.list();
         setCompetenceTypes(competenceTypes.body?.list ?? []);
-    }
-}
-
-type CompetenceTypeEditorProps = {
-    ct: CompetenceType;
-    refreshCTs: () => any;
-};
-
-function CompetenceTypeEditor(props: Readonly<CompetenceTypeEditorProps>) {
-    const { ct, refreshCTs } = props;
-
-    const [nameInput, setNameInput] = useState<string>();
-    const renderNameChange = nameInput !== undefined;
-
-    /* On click edit focus on the input */
-    useEffect(() => {
-        if (nameInput === undefined) return;
-
-        document
-            .querySelector(`.competence-editor[data-competence-name="${ct.name}"] form.competence-editing-form`)
-            ?.getElementsByTagName("input")
-            .item(0)?.focus();
-    }, [nameInput]);
-
-    const reviewTitle = ct.reviewed
-        ? "Essa competência já foi revisada por um administrador"
-        : "Clique para aprovar essa competência";
-
-    const reviewAction = ct.reviewed
-        ? () => {}
-        : reviewCompetence;
-
-    return <div className="competence-editor" data-competence-name={ct.name}>
-        { renderNameChange ||
-        <button disabled={ct.reviewed} title={reviewTitle} onClick={reviewAction} className="competence-interact-button">
-            <i className={makeClassName("bi", ct.reviewed ? "bi-check-circle-fill" : "bi-check-circle")} />
-        </button>
-        }
-
-        <h2>
-            { renderNameChange
-                ? <form className="competence-editing-form competencetype-name-wrapper">
-                    <button type="submit" className="competence-interact-button" onClick={saveNameChange} title="Salvar mudanças">
-                        <i className="bi bi-floppy-fill"/>
-                    </button>
-                    <button type="reset" className="competence-interact-button competence-cancel-name" onClick={() => setNameInput(undefined)} title="Cancelar mudanças">
-                        <i className="bi bi-x-circle"/>
-                    </button>
-                    <input type="text" className="competence-name-input competencetype-name" defaultValue={ct.name} onChange={setStateAsValue(setNameInput)} />
-                </form>
-                : <div className="competencetype-name-wrapper">
-                    <button type="button" className="competence-interact-button" onClick={() => setNameInput(ct.name)} title="Alterar nome da competência">
-                        <i className="bi bi-pencil-square"/>
-                    </button>
-                    <h3 className="competencetype-name"> { ct.name } </h3>
-                </div>
-            }
-        </h2>
-
-        <button type="button" className="competence-interact-button competencetype-delete" onClick={deleteCompetenceType} title="Excluir competência">
-            <i className="bi bi-trash-fill"/>
-        </button>
-    </div>
-
-    async function reviewCompetence() {
-        const response = await SwalUtils.fireModal({
-            title: `Deseja aprovar a competência "${ct.name}"?`,
-            text: "Essa competência se tornará visível para todos os usuários do Universi.me.",
-
-            showConfirmButton: true,
-            confirmButtonText: "Aprovar competência",
-            showCancelButton: true,
-            cancelButtonText: "Cancelar ação"
-        });
-
-        if (!response.isConfirmed)
-            return;
-
-        await UniversimeApi.CompetenceType.update({ id: ct.id, reviewed: true });
-        await refreshCTs();
-    }
-
-    async function saveNameChange(e: MouseEvent<HTMLButtonElement>) {
-        e.preventDefault();
-
-        if (nameInput === undefined)
-            return;
-
-        await UniversimeApi.CompetenceType.update({ id: ct.id, name: nameInput });
-        await refreshCTs();
-        setNameInput(undefined);
-    }
-
-    async function deleteCompetenceType() {
-        const response = await SwalUtils.fireModal({
-            title: `Deseja excluir a competência "${ct.name}"?`,
-            text: "Essa ação não pode ser desfeita",
-
-            showConfirmButton: true,
-            confirmButtonText: "Excluir",
-            showCancelButton: true,
-            cancelButtonText: "Cancelar",
-        });
-
-        if (!response.isConfirmed)
-            return;
-
-        await UniversimeApi.CompetenceType.remove({ id: ct.id });
-        await refreshCTs();
     }
 }
