+++ src/pages/Settings/CompetencesPage/index.ts
@@ -1,2 +1,3 @@
 export * from "./CompetencesSettingsLoader";
 export * from "./CompetencesSettingsPage";
+export * from "./CompetenceTypeEditor";
